@@ -15,8 +15,8 @@
 /// *** Code    *** *** *** *** *** *** *** *** *** *** *** *** *** *** *** ///
 package com.coffeine.virtuoso.module.user.model.entity;
 
-import org.codehaus.jackson.annotate.JsonIgnore;
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.NotEmpty;
 
@@ -63,7 +63,7 @@ public class Poet implements Serializable {
     @JoinColumn( name = "id_user" )
     protected User user;
 
-    @JsonIgnore
+    @JsonManagedReference
 //    @NotNull
 //    @NotEmpty
 //    @Valid
