@@ -8,8 +8,11 @@
 
 package com.coffeine.virtuoso.module.security.view.form;
 
+import com.coffeine.virtuoso.library.validator.anotation.Event;
 import com.coffeine.virtuoso.library.validator.anotation.InEnum;
 import com.coffeine.virtuoso.module.security.model.entity.Roles;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
 import org.hibernate.validator.constraints.Email;
 import org.hibernate.validator.constraints.NotEmpty;
 
@@ -23,6 +26,7 @@ import java.util.List;
  *
  * @version 1.0
  */
+@Event( start = "birthday", end = "deathDay" )
 public class RegistrationForm {
 
     /// *** Properties  *** ///
@@ -58,8 +62,10 @@ public class RegistrationForm {
     protected String locale;
 
     @NotNull
-    protected LocalDate birthday;
+    @JsonDeserialize( using = LocalDateDeserializer.class )
+    protected LocalDate birthday;//FIXME: Add validation for date format
 
+    @JsonDeserialize( using = LocalDateDeserializer.class )
     protected LocalDate deathDay;
 
 
